refactoring queue (#5)
1. Mutex controls locks of performable jobs in database.
2. Plan as a series of parameters structure a new job and support batch insert.
3. RetryInPlan retries according to RetryPolicy of Plan.
4. Args as []byte and give a Args() helper method.
5. RetryPolicy as per job instance level.
6. Mutex as goruntimes safe.

+++ retry_test.go
@@ -10,7 +10,7 @@ var intervalCases = []struct {
 	retryPolicy RetryPolicy
 	min         []time.Duration
 	max         []time.Duration
-	retryCount  int
+	retryCount  int32
 }{
 	{
 		name: "default",
@@ -54,9 +54,9 @@ func TestNextInterval(t *testing.T) {
 	for i := 0; i < 100; i++ {
 		for _, tc := range intervalCases {
 			policy := tc.retryPolicy
-			var retryCount int
+			var retryCount int32
 			for i := range tc.min {
-				interval, ok := policy.NextInterval(i)
+				interval, ok := policy.NextInterval(int32(i))
 				if !ok {
 					t.Fatalf("%s(%v): want interval in [%v, %v] but retry finished", tc.name, i, tc.min[i], tc.max[i])
 				}
